Relocate server autoload code to a new module
